backend/requirements.txt
fastapi
uvicorn
pydantic
torch
torchvision
python-multipart
pillow
requests
httpx
python-dotenv
fastapi[all]
